pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nathreader"
version = "0.1.0"
description = "A modern PDF and document reader application"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nath", email = "nath@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyMuPDF>=1.19.0",
    "python-docx>=0.8.11",
    "python-pptx>=0.6.21",
    "Pillow>=9.0.0",
    "customtkinter>=5.2.0",
    "darkdetect"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.12b0",
    "isort>=5.10.1",
    "mypy>=0.910",
    "flake8>=4.0.1"
]

[project.scripts]
nathreader = "nathreader.__main__:main"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
